Add JDBC API and H2 Database dependencies
* Added the 'spring-boot-starter-jdbc' dependency to the project's
dependencies in the pom.xml file. The Spring Boot starter for JDBC (Java
Database Connectivity) includes all the necessary libraries and
configurations for working with relational databases that support the
JDBC API, such as H2, MySQL, PostgreSQL, etc, in a Spring Boot
application
* Added the 'h2' dependency to the project's dependency in the pom.xml
file. The H2 database is good for development and testing purposes due
to its simplicity and ease of use. It's particularly useful for writing
and running unit tests because it's lightweight and can be embedded
within your application. The 'runtime' scope means that this dependency
is only necessary at runtime, not for compiling your code